+++ CHANGELOG.md
@@ -55,8 +55,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Bump the SDK to `v0.50.5-pio-1` (from an earlier ephemeral version) [#1897](https://github.com/provenance-io/provenance/pull/1897).
 * Removed `rewards` module [#1905](https://github.com/provenance-io/provenance/pull/1905).
 * Remove unused navs [#1920](https://github.com/provenance-io/provenance/issues/1920).
+* Update genutil for sdk 50 [#1760](https://github.com/provenance-io/provenance/issues/1760).
 * Migrate module params from param space to module store.
   *  Attribute module param migration [#1927](https://github.com/provenance-io/provenance/pull/1927)
+  *  Metadata module param migration [#1932](https://github.com/provenance-io/provenance/pull/1932)
 
 ### Dependencies